Refactor env to registry package
Env package now contains one type that is specific to the environment.
The registry package contains changes which are specific to retrieving
the remote ws urls. So this commit splits the two different concerns.
A future commit with also move the remote registry specific logic from
the browser package into the registry package.
