+++ registry/doc.go
@@ -1,0 +1,3 @@
+// Package registry provides types and methods that store
+// information that currently span the whole test run.
+package registry
